requirements.txt
async-timeout==4.0.3
asyncpg==0.29.0
greenlet==3.0.3
pyaes==1.6.1
Pyrogram @ git+https://github.com/dyanashek/pyrogram@a086b492039687dd1b807969f9202061ce5305da
PySocks==1.7.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
TgCrypto==1.2.5
typing_extensions==4.11.0
uvloop==0.19.0
